# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/condominios": {
      "post": {
        "description": "Crea un nuevo condominio con su base de datos dedicada. Requiere rol SUPERADMIN.",
        "operationId": "CondominiosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateCondominioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Condominio creado exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "403": {
            "description": "No autorizado - Se requiere rol SUPERADMIN"
          }
        },
        "security": [
          {
            "better-auth.session_token": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Crear un nuevo condominio",
        "tags": [
          "condominios"
        ]
      },
      "get": {
        "description": "Retorna una lista de todos los condominios registrados",
        "operationId": "CondominiosController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de condominios obtenida exitosamente"
          }
        },
        "summary": "Obtener todos los condominios",
        "tags": [
          "condominios"
        ]
      }
    },
    "/condominios/users": {
      "post": {
        "description": "Crea un nuevo usuario en el condominio detectado del subdominio. Requiere rol SUPERADMIN o ADMIN.",
        "operationId": "CondominiosController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCondominioUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente"
          },
          "400": {
            "description": "Datos inválidos o subdominio no detectado"
          },
          "403": {
            "description": "No autorizado - Se requiere rol SUPERADMIN o ADMIN"
          }
        },
        "security": [
          {
            "better-auth.session_token": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Crear un nuevo usuario en el condominio",
        "tags": [
          "condominios"
        ]
      },
      "get": {
        "description": "Retorna una lista de todos los usuarios del condominio detectado del subdominio. Requiere rol SUPERADMIN o ADMIN.",
        "operationId": "CondominiosController_getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida exitosamente"
          },
          "403": {
            "description": "No autorizado - Se requiere rol SUPERADMIN o ADMIN"
          }
        },
        "security": [
          {
            "better-auth.session_token": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener todos los usuarios del condominio",
        "tags": [
          "condominios"
        ]
      }
    },
    "/condominios/users/{userId}": {
      "get": {
        "operationId": "CondominiosController_getUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "condominios"
        ]
      },
      "put": {
        "operationId": "CondominiosController_updateUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCondominioUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "condominios"
        ]
      },
      "patch": {
        "operationId": "CondominiosController_patchUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCondominioUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "condominios"
        ]
      },
      "delete": {
        "operationId": "CondominiosController_deleteUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "condominios"
        ]
      }
    },
    "/condominios/users/{userId}/role": {
      "patch": {
        "operationId": "CondominiosController_updateUserRole",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "condominios"
        ]
      }
    },
    "/condominios/{id}": {
      "get": {
        "operationId": "CondominiosController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "condominios"
        ]
      },
      "put": {
        "operationId": "CondominiosController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCondominioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "condominios"
        ]
      }
    },
    "/condominios/{id}/deactivate": {
      "post": {
        "operationId": "CondominiosController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "condominios"
        ]
      }
    },
    "/condominios/{id}/delete": {
      "post": {
        "operationId": "CondominiosController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "condominios"
        ]
      }
    },
    "/condominios/validate-subdomain/{subdomain}": {
      "get": {
        "operationId": "CondominiosController_validateSubdomain",
        "parameters": [
          {
            "name": "subdomain",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "condominios"
        ]
      }
    },
...
    "/condominios/login": {
      "post": {
        "description": "Autentica un usuario de condominio y establece una sesión. El condominio se detecta automáticamente del subdominio si no se proporciona condominioId.",
        "operationId": "CondominiosController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCondominioUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "user": {
                      "id": "uuid",
                      "email": "juan.perez@example.com",
                      "name": "Juan Pérez",
                      "role": "ADMIN"
                    },
                    "session": {
                      "token": "session_token",
                      "expiresAt": "2024-12-31T23:59:59.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        },
        "summary": "Iniciar sesión como usuario de condominio",
        "tags": [
          "condominios"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Vekino",
    "description": "API para gestión de condominios",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.vekino.com",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Token de autenticación JWT",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Cookie de sesión de Better Auth"
      }
    },
    "schemas": {
...
      "CreateCondominioDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del condominio",
            "example": "Condominio Las Flores",
            "minLength": 3
          },
          "nit": {
            "type": "string",
            "description": "NIT / Identificación Tributaria",
            "example": "900123456-7"
          },
          "address": {
            "type": "string",
            "description": "Dirección física del condominio",
            "example": "Calle 123 #45-67"
          },
          "city": {
            "type": "string",
            "description": "Ciudad donde se encuentra el condominio",
            "example": "Bogotá"
          },
          "country": {
            "type": "string",
            "description": "País donde se encuentra el condominio",
            "example": "Colombia"
          },
          "timezone": {
            "type": "string",
            "description": "Zona horaria del condominio",
            "enum": [
              "AMERICA_BOGOTA",
              "AMERICA_MEXICO_CITY",
              "AMERICA_LIMA",
              "AMERICA_SANTIAGO",
              "AMERICA_BUENOS_AIRES",
              "AMERICA_CARACAS",
              "AMERICA_MONTEVIDEO",
              "AMERICA_ASUNCION",
              "AMERICA_LA_PAZ",
              "AMERICA_QUITO",
              "AMERICA_GUAYAQUIL",
              "AMERICA_PANAMA",
              "AMERICA_MANAGUA",
              "AMERICA_SAN_JOSE",
              "AMERICA_TEGUCIGALPA",
              "AMERICA_GUATEMALA",
              "AMERICA_SANTO_DOMINGO",
              "AMERICA_HAVANA",
              "UTC"
            ],
            "example": "AMERICA_BOGOTA"
          },
          "frontSubdomain": {
            "type": "string",
            "description": "Subdominio usado por el frontend (ej: las-flores). Solo letras minúsculas, números y guiones",
            "example": "las-flores",
            "pattern": "^[a-z0-9-]+$"
          },
          "subdomain": {
            "type": "string",
            "description": "Subdominio único para acceso (opcional, se genera automáticamente desde frontSubdomain)",
            "example": "condominio1",
            "pattern": "^[a-z0-9-]+$"
          },
          "logo": {
            "type": "string",
            "description": "URL del logo (se genera automáticamente al subir imagen)",
            "example": "https://example.com/logo.png"
          },
          "primaryColor": {
            "type": "string",
            "description": "Color principal en formato hexadecimal",
            "example": "#3B82F6",
            "pattern": "^#[0-9A-Fa-f]{6}$"
          },
          "subscriptionPlan": {
            "type": "string",
            "description": "Plan de suscripción del condominio",
            "enum": [
              "BASICO",
              "PRO",
              "ENTERPRISE"
            ],
            "example": "BASICO"
          },
          "unitLimit": {
            "type": "number",
            "description": "Límite de unidades permitidas",
            "example": 100,
            "minimum": 1
          },
          "planExpiresAt": {
            "type": "string",
            "description": "Fecha de vencimiento del plan (ISO 8601)",
            "example": "2024-12-31T23:59:59.000Z"
          },
          "activeModules": {
            "description": "Módulos activos del condominio",
            "example": [
              "reservas",
              "documentos",
              "pqrs"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "databaseName": {
            "type": "string",
            "description": "Nombre de la base de datos (opcional, se genera automáticamente desde el nombre)",
            "example": "condominio_las_flores"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateCondominioUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre completo del usuario",
            "example": "Juan Pérez"
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico del usuario",
            "example": "juan.perez@example.com"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario (mínimo 6 caracteres)",
            "example": "password123",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario en el condominio",
            "enum": [
              "ADMIN",
              "PROPIETARIO",
              "ARRENDATARIO",
              "RESIDENTE"
            ],
            "example": "ADMIN"
          },
          "firstName": {
            "type": "string",
            "description": "Primer nombre (requerido para PROPIETARIO, ARRENDATARIO, RESIDENTE)",
            "example": "Juan"
          },
          "lastName": {
            "type": "string",
            "description": "Apellido (requerido para PROPIETARIO, ARRENDATARIO, RESIDENTE)",
            "example": "Pérez"
          },
          "tipoDocumento": {
            "type": "string",
            "description": "Tipo de documento de identidad",
            "example": "CC"
          },
          "numeroDocumento": {
            "type": "string",
            "description": "Número de documento de identidad",
            "example": "1234567890"
          },
          "telefono": {
            "type": "string",
            "description": "Número de teléfono",
            "example": "+57 300 123 4567"
          },
          "unidadId": {
            "type": "string",
            "description": "ID de la unidad asociada al usuario",
            "example": "uuid"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "UpdateCondominioUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre completo del usuario",
            "example": "Juan Pérez"
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico del usuario",
            "example": "juan.perez@example.com"
          },
          "password": {
            "type": "string",
            "description": "Nueva contraseña (mínimo 6 caracteres)",
            "example": "newpassword123",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario en el condominio",
            "enum": [
              "ADMIN",
              "PROPIETARIO",
              "ARRENDATARIO",
              "RESIDENTE"
            ],
            "example": "ADMIN"
          },
          "firstName": {
            "type": "string",
            "description": "Primer nombre",
            "example": "Juan"
          },
          "lastName": {
            "type": "string",
            "description": "Apellido",
            "example": "Pérez"
          },
          "tipoDocumento": {
            "type": "string",
            "description": "Tipo de documento de identidad",
            "example": "CC"
          },
          "numeroDocumento": {
            "type": "string",
            "description": "Número de documento de identidad",
            "example": "1234567890"
          },
          "telefono": {
            "type": "string",
            "description": "Número de teléfono",
            "example": "+57 300 123 4567"
          },
          "unidadId": {
            "type": "string",
            "description": "ID de la unidad asociada al usuario",
            "example": "uuid"
          }
        }
      },
      "UpdateCondominioDto": {
        "type": "object",
        "properties": {}
      },
      "LoginCondominioUserDto": {
        "type": "object",
        "properties": {
          "condominioId": {
            "type": "string",
            "description": "ID del condominio (opcional, se detecta del subdominio si no se proporciona)",
            "example": "uuid"
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico del usuario",
            "example": "juan.perez@example.com"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}